+++ js/API_URLs_export.mjs
@@ -5,5 +5,9 @@
  const signinrUrl = `${baseUrl_API}/api/v1/social/auth/login`;
  const posts = `${baseUrl_API}/api/v1/social/posts/?_author=true`;
 
+
+
 export {signinrUrl, registerUrl, posts, authorization};
+
+
