+++ js/post_feed_profile.mjs
@@ -1,4 +1,4 @@
-import { posts, authorization} from "./API_URLs.mjs";
+import { posts, authorization} from "./API_URLs_export.mjs";
 import {displayPostsProfile} from "./post_feed_export.mjs";
