+++ js/signin.mjs
@@ -1,12 +1,7 @@
 
-import { signinrUrl } from "./API_URLs.mjs";
+import { signinrUrl} from "./API_URLs_export.mjs";
+import {email, password, displayMessage, submit, checkIfLoggedIn} from "./register_signin_export.mjs";
 
-const email = document.querySelector("#floatingInput");
-const password = document.querySelector("#floatingPassword");
-const displayMessage = document.querySelector(".h1");
-const removeForm = document.querySelectorAll(".form-floating");
-const submit = document.querySelector("#submit");
-submit.addEventListener("click", valueInput);
 
 function valueInput (event) {
     event.preventDefault();
@@ -36,7 +31,7 @@ function valueInput (event) {
                 localStorage.setItem("userName", json.name)
                 window.open("profile.html", "_self");
             } else {
-                displayMessage.innerHTML = `wrong email or password, try again`;
+                displayMessage.innerHTML =  `invalid email or password please try again`;
             }
             
     
@@ -48,21 +43,11 @@ function valueInput (event) {
   
   registerUser(signinrUrl, userInput) 
 }    
-    
 
-const checkIfLoggedIn = () => {
-    const getLocal = localStorage.getItem("accsessToken");
-if (getLocal !== null){
-    removeForm.forEach((e) => {
-        e.style.display = "none";
-    });
-    displayMessage.innerHTML = `You are logged in`;
-    submit.style.display = "none";
-}
-
-}
+submit.addEventListener("click", valueInput);
 
 checkIfLoggedIn();
+
