+++ profile.html
@@ -97,7 +97,7 @@
     </div>
     <script src="/node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
     <script type="module" src="/js/logout.mjs"></script>
-    <script type="module" src="/js/post_feed.mjs"></script>
+    <script type="module" src="/js/post_feed_profile.mjs"></script>
     <script type="module" src="/js/profile.mjs"></script>    
 </body>
 </html>
